moved buy functions from loops/user.rs to loops/management.rs
